fix: 더보기 관련 오류 수정
- match를 각자 GameSlot에서 recoil로 관리하도록 수정
