+++ client/pages/summoners/[summonerName].tsx
@@ -21,8 +21,8 @@ import {
   tryToGetSummonerProfile,
 } from '../../utils/api';
 import { getUsersFromLocalStorage, setUsersToLocalStorage } from '../../utils/localStorage';
-import { useRecoilState } from 'recoil';
-import { matchStatisticState } from '../../states/gameSlot';
+import { useRecoilState, useRecoilValue } from 'recoil';
+import { matchStateFamily, matchStatisticState } from '../../states/gameSlot';
 
 interface SummonerProfilePanelProps {
   summonerName: string;
@@ -165,13 +165,12 @@ export default function SearchPage({ name }: SearchPageProps) {
 function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerProfilePanelProps) {
   const [summonerProfile, setSummonerProfile] = useState<SummonerProfile | null>(null);
   const [matchIds, setMatchIds] = useState<SummonerMatchIds>([]);
-  const [matches, setMatches] = useState<Match[]>([]);
   const [fetching, setFetching] = useState<boolean>(false);
   const [pollStart, setPollStart] = useState<Date | null>(null);
   const [loadMore, setLoadMore] = useState<number | null>(null);
   const [update, setUpdate] = useState<number | null>(null);
-  const [scoreMultipliers, setScoreMultipliers] = useState<Contribution | null>(null);
   const [, setMatchStatistics] = useRecoilState(matchStatisticState);
+  const lastMatch = useRecoilValue(matchStateFamily(matchIds[matchIds.length - 1]));
 
   // tick for load more
   useEffect(() => {
@@ -179,14 +178,16 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
       const newSummonerProfile = await getSummonerProfile(summonerName);
       if (pollStart != newSummonerProfile.updatedAt) {
         setSummonerProfile(newSummonerProfile);
-        const lastMatchId = matches.length ? matches[matches.length - 1].info.gameId : 0;
+        const lastMatchId = matchIds.length ? matchIds[matchIds.length - 1] : 0;
 
         const newMatchIds = (
           await getSummonerMatchIds(newSummonerProfile.puuid, lastMatchId)
         ).matchIds.filter((id) => matchIds.indexOf(id) === -1);
 
-        setMatchIds(newMatchIds);
+        setMatchIds((prev) => [...prev, ...newMatchIds]);
         setLoadMore(null);
+        setFetching(false);
+        setPollStart(null);
       } else if (loadMore !== null) setLoadMore(loadMore + 1);
     }
 
@@ -194,7 +195,7 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
       const timer = setTimeout(tick, 1000);
       return () => clearTimeout(timer);
     }
-  }, [loadMore, summonerName, matchIds, matches, pollStart]);
+  }, [loadMore, summonerName, matchIds, pollStart]);
 
   // tick for update user
   useEffect(() => {
@@ -202,7 +203,6 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
       const newSummonerProfile = await getSummonerProfile(summonerName);
       if (pollStart != newSummonerProfile.updatedAt) {
         setSummonerProfile(newSummonerProfile);
-        setMatches([]);
         setMatchIds((await getSummonerMatchIds(newSummonerProfile.puuid)).matchIds);
         setUpdate(null);
       } else if (update !== null) setUpdate(update + 1);
@@ -223,7 +223,6 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
     setPollStart(null);
     setSummonerProfile(null);
     setMatchIds([]);
-    setMatches([]);
     setMatchStatistics({});
 
     if (!summonerName) return;
@@ -239,29 +238,7 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
         }
       }
     })();
-
-    (async () => {
-      setScoreMultipliers(await getScoreMultipliers());
-    })();
   }, [setMatchStatistics, setSummonerNotFound, summonerName]);
-
-  useEffect(() => {
-    if (matchIds.length === 0 || !scoreMultipliers) return;
-
-    const promises = matchIds.map(async (matchId) => {
-      const match = await getMatch(matchId);
-      return match;
-    });
-
-    (async () => {
-      try {
-        const newMatches = await Promise.all(promises);
-        setMatches((matches) => [...matches, ...newMatches]);
-      } catch (e) {
-        console.error(e);
-      }
-    })();
-  }, [matchIds, scoreMultipliers]);
 
   // before loading router
   if (!summonerName) return <></>;
@@ -292,6 +269,7 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
   };
 
   const onLoadMore = () => {
+    if (!lastMatch) return;
     setFetching(true);
     (async () => {
       try {
@@ -299,7 +277,7 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
           (
             await requestFetchSummonerMatches(
               summonerProfile.puuid,
-              Math.floor(matches[matches.length - 1].info.gameCreation / 1000),
+              Math.floor(lastMatch.match.info.gameCreation / 1000),
             )
           ).startedAt,
         );
@@ -311,7 +289,7 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
 
       try {
         // get matches from local (fallback)
-        const lastMatchId = matches.length ? matches[matches.length - 1].info.gameId : 0;
+        const lastMatchId = matchIds.length ? matchIds[matchIds.length - 1] : 0;
         const newMatchIds = (
           await getSummonerMatchIds(summonerProfile.puuid, lastMatchId)
         ).matchIds.filter((id) => matchIds.indexOf(id) === -1);
@@ -346,12 +324,8 @@ function SummonerProfilePanel({ summonerName, setSummonerNotFound }: SummonerPro
         <SummonerStatCard />
       </div>
       <main css={style.main}>
-        {matchIds.map((matchId, idx) => (
-          <GameSlot
-            key={`gameSlot-${matchId}-${idx}`}
-            puuid={summonerProfile.puuid}
-            matchId={matchId}
-          />
+        {matchIds.map((matchId) => (
+          <GameSlot key={`gameSlot-${matchId}`} puuid={summonerProfile.puuid} matchId={matchId} />
         ))}
         {matchIds.length !== 0 && (
           <Button width={'100%'} enabled={!fetching} onClick={onLoadMore}>
